rename classes & move resend stuff to foler
