+++ src/NetworkNode.js
@@ -1,6 +1,6 @@
 const { ControlLayer, MessageLayer } = require('streamr-client-protocol')
 
-const { StorageResendStrategy, StorageNodeResendStrategy } = require('./logic/resendStrategies')
+const { LocalResendStrategy, ForeignResendStrategy } = require('./resend/resendStrategies')
 const Node = require('./logic/Node')
 const { StreamIdAndPartition } = require('./identifiers')
 
@@ -12,8 +12,8 @@ class NetworkNode extends Node {
         const networkOpts = {
             ...opts,
             resendStrategies: [
-                ...opts.storages.map((storage) => new StorageResendStrategy(storage)),
-                new StorageNodeResendStrategy(
+                ...opts.storages.map((storage) => new LocalResendStrategy(storage)),
+                new ForeignResendStrategy(
                     opts.protocols.trackerNode,
                     opts.protocols.nodeToNode,
                     (streamKey) => this._getTrackerId(streamKey),
